server side changes to include module restructuring
