+++ server/src/main/java/com/flipkart/varadhi/authn/CustomAuthenticationHandler.java
@@ -4,10 +4,10 @@ import com.flipkart.varadhi.config.AuthenticationConfig;
 import com.flipkart.varadhi.entities.Org;
 import com.flipkart.varadhi.entities.auth.UserContext;
 import com.flipkart.varadhi.common.exceptions.InvalidConfigException;
-import com.flipkart.varadhi.spi.RequestContext;
-import com.flipkart.varadhi.spi.authn.AuthenticationHandlerProvider;
-import com.flipkart.varadhi.spi.authn.Authenticator;
-import com.flipkart.varadhi.spi.utils.OrgResolver;
+import com.flipkart.varadhi.server.spi.RequestContext;
+import com.flipkart.varadhi.server.spi.authn.AuthenticationHandlerProvider;
+import com.flipkart.varadhi.server.spi.authn.Authenticator;
+import com.flipkart.varadhi.server.spi.utils.OrgResolver;
 import io.micrometer.core.instrument.MeterRegistry;
 import io.vertx.core.Future;
 import io.vertx.core.Vertx;
