+++ server/src/main/java/com/flipkart/varadhi/web/AuthzHandler.java
@@ -1,8 +1,8 @@
 package com.flipkart.varadhi.web;
 
 import com.flipkart.varadhi.spi.ConfigFileResolver;
-import com.flipkart.varadhi.spi.authz.AuthorizationOptions;
-import com.flipkart.varadhi.spi.authz.AuthorizationProvider;
+import com.flipkart.varadhi.server.spi.authz.AuthorizationOptions;
+import com.flipkart.varadhi.server.spi.authz.AuthorizationProvider;
 import com.flipkart.varadhi.config.AppConfiguration;
 import com.flipkart.varadhi.common.exceptions.InvalidConfigException;
 import com.flipkart.varadhi.web.routes.RouteConfigurator;
